claude_a2a/pyproject.toml
[project]
name = "claude-a2a"
version = "0.1.0"
description = "A2A protocol server for Claude Code VPS agent"
requires-python = ">=3.12"
dependencies = [
    "a2a-sdk>=0.2.0",
    "httpx>=0.27.0",
    "uvicorn>=0.34.0",
    "starlette>=0.45.0",
]

[project.optional-dependencies]
dev = ["pytest", "ruff"]
